#58 Refactored gspread out of data
