+++ shopping_list/data.py
@@ -1,10 +1,7 @@
 from .database import item_database
 from .database import recipe_database
 
-import gspread
-from oauth2client.service_account import ServiceAccountCredentials
-
-items_sheet_name = 'Items'
+ingredients_sheet_name = 'Items'
 recipes_sheet_name = 'Recipes'
 input_sheet_name = 'Input'
 
@@ -24,9 +21,8 @@ class Data:
         '''
         Pull data from the ingredients sheet into a list of dicts.
         '''
-        self._ingredients_sheet_data = spreadsheet.worksheet(items_sheet_name).get_all_records()
+        self._ingredients_sheet_data = spreadsheet.worksheet(ingredients_sheet_name).get_all_records()
         print('ingredients retrieved')
-        return
 
     def download_recipes_data(self, spreadsheet):
         '''
@@ -34,28 +30,26 @@ class Data:
         '''
         self._recipes_sheet_data = spreadsheet.worksheet(recipes_sheet_name).get_values()
         print('recipes retrieved')
-        return
 
     def download_input_data(self, spreadsheet):
         '''
         Pull data from the inputs sheet into a list of lists.
         '''
-        self.input_list = spreadsheet.worksheet(input_sheet_name).get_values(major_dimension="COLUMNS")
-        # Get input config data.
+        input_sheet_data_raw = spreadsheet.worksheet(input_sheet_name).get_values(major_dimension="COLUMNS")
         self._input_sheet_data = {}
-        for column in self.input_list:
+        for column in input_sheet_data_raw:
             column_heading = column[0]
             # Remove any empty strings with list comprehension.
             column_data = list(filter(None, column[1:]))
             self._input_sheet_data[column_heading] = column_data
-        print('input retrieved')
-        return
 
-    def add_new_recipe_to_buy(self, recipes_to_buy_list, new_recipe_name):
+        print('input retrieved')
+
+    def add_new_recipe_to_buy(self, spreadsheet, new_recipe_name):
+        recipes_to_buy_list = self._input_sheet_data['Meals To Buy']
         new_recipe_row = len(recipes_to_buy_list) + 1
-        # TODO magic number
         new_recipe_col = 1
-        self._input_sheet.update_cell(new_recipe_row, new_recipe_col, new_recipe_name)
+        spreadsheet.worksheet(input_sheet_name).update_cell(new_recipe_row, new_recipe_col, new_recipe_name)
 
     def get_ingredient_list(self):
         return [x['Name'] for x in self._ingredients_sheet_data]
@@ -65,7 +59,9 @@ class Data:
         return grouping_options
 
     def get_ingredient_sheet_data(self, grouping_selection):
-        # Construct the ingredients database.
+        '''
+        Construct the ingredients database.
+        '''
         ingredients = item_database.ItemDatabase()
         for record in self._ingredients_sheet_data:
             # The record holds the name of the ingredient and all of the groupings. The name and the groupings must be
@@ -81,7 +77,9 @@ class Data:
         return ingredients
 
     def get_recipe_sheet_data(self):
-        # Construct recipes database.
+        '''
+        Construct recipes database.
+        '''
         recipes = recipe_database.RecipeDatabase()
         for recipe_row in self._recipes_sheet_data:
             # The row holds the name of the recipe in the first column and all of the ingredients in subsequent columns.
@@ -93,26 +91,11 @@ class Data:
         return recipes
 
     def get_input_sheet_data(self):
+        '''
+        Get input config data.
+        '''
         recipes_to_buy_list = self._input_sheet_data['Meals To Buy']
         exclusions_list = self._input_sheet_data['Exclusions']
         inclusions_list = self._input_sheet_data['Inclusions']
 
         return recipes_to_buy_list, exclusions_list, inclusions_list
-
-
-def open_spreadsheet(token_filename, sheet_name):
-    # use creds to create a client to interact with the Google Drive API
-    scope = [
-        'https://spreadsheets.google.com/feeds',
-        'https://www.googleapis.com/auth/drive'
-    ]
-    # TODO put filename in config file
-    creds = ServiceAccountCredentials.from_json_keyfile_name(
-        token_filename, scope)
-    client = gspread.authorize(creds)
-
-
-    # Find a spreadsheet by name and open it.
-    spreadsheet = client.open(sheet_name)
-    print('data connected')
-    return spreadsheet
